requirements.txt
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
ccxt==4.2.25
yfinance==0.2.36
alpha-vantage==2.3.1
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
joblib==1.3.2
nltk==3.8.1
vaderSentiment==3.3.2
python-telegram-bot==20.7
jinja2==3.1.3
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
colorlog==6.8.2